+++ metadata/changelog.md
@@ -1,9 +1,29 @@
 ---
 layout: page
 title: Changelog
-last_modified_at: 2026-01-01
+last_modified_at: 2026-01-18
 permalink: /changelog
 ---
+
+## 2026-01-18
+
+### Added
+
+- [Admin dashboard from Odin Project](/odinProject/adminDashboard.html)
+- [Battleship from Odin Project](/odinProject/battleship.html)
+- [Calculator from Odin Project](/odinProject/calculator.html)
+- [Home Page from Odin Project](/odinProject/homePage.html)
+- [Library from Odin Project](/odinProject/library.html)
+- [Recipes from Odin Project](/odinProject/recipes.html)
+- [Product Landing Page from Odin Project](/odinProject/productLandingPage.html)
+- [Restaurant Page from Odin Project](/odinProject/restaurantPage/index.html)
+- [Signup page from Odin Project](/odinProject/signup.html)
+- [TODOs page from Odin Project](/odinProject/todos.html)
+- [Moved index page](/odinProject/)
+
+## Modified
+
+- Updated NavBar
 
 ## 2026-01-01